+++ database/migrations/2021_09_09_075910_create_task_users_table.php
@@ -15,8 +15,8 @@ class CreateTaskUsersTable extends Migration
     {
         Schema::create('task_users', function (Blueprint $table) {
             $table->id();
-            $table->integer('task_id');
-            $table->integer('user_id');
+            $table->bigInteger('task_id');
+            $table->bigInteger('user_id');
             $table->timestamps();
             $table->softDeletes();
         });
