+++ database/migrations/2021_09_21_091034_create_group_conversations_table.php
@@ -15,11 +15,11 @@ class CreateGroupConversationsTable extends Migration
     {
         Schema::create('group_conversations', function (Blueprint $table) {
             $table->id();
-            $table->integer('project_id');
-            $table->integer('user_id');
+            $table->bigInteger('project_id');
+            $table->bigInteger('user_id');
             $table->text('message')->nullable();
             $table->integer('message_type')->nullable();
-            $table->integer('document_id')->nullable();
+            $table->bigInteger('document_id')->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
